+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [2.3.6] - (04.01.2023)
+## Changes
+- fix save target visibility
+- fix save custom names 
+
 ## [2.3.3] - (31.12.2022)
 ## Changes
 - dynamic update accessory information